crimson/os/seastore: introduce NVMeIODriver and decouple NVMeBlockDevice
Extract the raw NVMe control/data plane (pass_admin, pass_through_io, get_nsid,
identify_*) out of NVMeBlockDevice into an NVMeIODriver interface with a
KernelNVMeIODriver implementation, and move the NVMe command structs into a new
nvme_command.h so the RBM device no longer depends on linux/nvme_ioctl.h. PI
state is pushed in via set_protection_info().

While moving the kernel backend, fix pass_admin(): it passed nullptr to
ioctl(NVME_IOCTL_ADMIN_CMD) instead of &admin_cmd, so admin/identify on the
kernel path could never reach the controller.
