+++ src/crimson/os/seastore/CMakeLists.txt
@@ -46,6 +46,7 @@ set(crimson_seastore_srcs
   random_block_manager.cc
   random_block_manager/block_rb_manager.cc
   random_block_manager/rbm_device.cc
+  random_block_manager/nvme_io_driver.cc
   random_block_manager/nvme_block_device.cc
   random_block_manager/hdd_device.cc
   random_block_manager/avlallocator.cc
